hm_12/pyproject.toml
[tool.poetry]
name = "hm_12"
version = "0.1.0"
description = ""
authors = ["Yaroslav Melnychuk <yarmel.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
psycopg2-binary = "^2.9.7"
sqlalchemy = "^2.0.20"
faker = "^19.6.0"
requests = "^2.31.0"
alembic = "^1.12.0"
libgravatar = "^1.0.4"
jose = "^1.0.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
